refactor(condo): moved input form into step
